pyproject.toml
[project]
name = "pulse"
version = "0.1.0"
description = "RSS Signal Generator for Sentinel"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "feedparser>=6.0.0",
    "readability-lxml>=0.8.0",
    "transformers>=4.48.0",
    "torch>=2.0.0",
    "gliclass>=0.1.0",
    "apscheduler>=3.10.0",
    "jinja2>=3.1.0",
    "sse-starlette>=2.0.0",
    "python-multipart>=0.0.9",
    "lxml-html-clean>=0.4.0",
    "psutil>=5.9.0",
]

[project.scripts]
pulse = "pulse.main:main"

[tool.setuptools.packages.find]
include = ["pulse*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
